requirements.txt
# Dependencias principales
google-generativeai>=0.3.0
pydantic>=2.0.0
python-dateutil>=2.8.0
customtkinter>=5.2.0

# Dependencias de desarrollo y testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Dependencias opcionales para mejoras
pillow>=9.0.0  # Para mejor soporte de imágenes en customtkinter
requests>=2.28.0  # Para futuras funcionalidades de red